token-monitor/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
httpx==0.28.1
apscheduler==3.11.0
python-dotenv==1.1.0
jinja2==3.1.6
